Address PR comments and adds the tests for ClusterAPI
